maint: avoid warning for printing pid_t on Solaris 11
The type pid_t is defined as long on Solaris 11.  Therefore, GCC
issued a -Wformat warning there when attempting to print as %d.

* find/exec.c: (launch): When writing the debug message, treat child_pid
as the wider long type.

+++ find/exec.c
@@ -373,9 +373,11 @@ launch (struct buildcmd_control *ctl, void *usercontext, int argc, char **argv)
   int ex = WEXITSTATUS (execp->last_child_status);
   if (options.debug_options & DebugExec)
     {
+      /* pid_t is of type long on Solaris 11.  Cast CHILD_PID for use with
+       * %ld as long as gnulib doesn't provide portable PRIdPID. */
       fprintf (stderr,
-               "DebugExec: process (PID=%d) terminated with exit status: %d\n",
-               child_pid, ex);
+               "DebugExec: process (PID=%ld) terminated with exit status: %d\n",
+               (long) child_pid, ex);
     }
 
   if (0 == ex)
